requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic~=2.11.5
python-dotenv==1.0.0
sqlalchemy==2.0.22
alembic==1.12.0
pytest==7.4.2
httpx==0.25.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6


langchain~=0.3.25
h11~=0.14.0
future~=0.18.2
pip~=24.3.1
wheel~=0.37.0
requests~=2.32.3
PyYAML~=6.0.2
exceptiongroup~=1.3.0
sniffio~=1.3.1
click~=8.1.8
httpcore~=0.18.0
idna~=3.10
zstandard~=0.23.0
charset-normalizer~=3.4.2
setuptools~=44.1.1
packaging~=24.2
urllib3~=2.4.0
MarkupSafe~=3.0.2
orjson~=3.10.18
certifi~=2025.4.26
six~=1.15.0
anyio~=3.7.1
annotated-types~=0.7.0
tenacity~=9.1.2
langchain-core~=0.3.64
langsmith~=0.3.45
langchain-text-splitters~=0.3.8
async-timeout~=4.0.3
requests-toolbelt~=1.0.0
jsonpointer~=3.0.0
starlette~=0.27.0
jsonpatch~=1.33
streamlit~=1.45.1
GitPython~=3.1.44
gitdb~=4.0.12
numpy~=2.0.2
pillow~=11.2.1
attrs~=25.3.0
protobuf~=6.31.1
pytz~=2025.2
toml~=0.10.2
smmap~=5.0.2
altair~=5.5.0
jsonschema~=4.24.0
narwhals~=1.41.1
pandas~=2.3.0
pyarrow~=20.0.0
jinja2~=3.1.6
referencing~=0.36.2
aiohttp~=3.12.9
yarl~=1.20.0
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
tzdata~=2025.2
tornado~=6.5.1
langchain-community>=0.3.24
cachetools~=5.5.2
blinker~=1.9.0
jsonschema-specifications~=2025.4.1